requirements.txt
dash==3.0.4
dash-ag-grid==2.4.0
dash-bootstrap-components==1.7.1
dash_mantine_components==1.3.0
dash-leaflet==1.1.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-Session==0.4.0
pandas==2.2.3
plotly==5.13.0
python-dotenv==0.19.1
requests==2.32.3
dash-iconify==0.1.2
gunicorn==23.0.0
werkzeug==3.0.6
stravalib==2.3
polyline==2.0.2
pymongo==4.8.0